The transaction history of a player can now be seen when clicked on a button

+++ Server/Services/TransferModels/Responses/PlayerDTO.cs
@@ -3,7 +3,7 @@
 public class PlayerDTO
 {
     public Guid PlayerId { get; set; }
-    public decimal? Balance { get; set; }
+    public int? Balance { get; set; }
     public string UserName { get; set; }
     public string Email { get; set; }
 }
